refactor: 이름, 구조 변경
- result는 사용하지 않는 무의미한 구조라고 판단하여 Custom Decodable 구조체 선언
- DailyBoxOffice 구조체 역시 DailyBoxOfficeSearchResult내에서만 사용하므로 nested struct로 변경
